Always get a default value for the types argument
When finishline calls Mosaic, it passes in an args dict that bypasses
the default argument parsing that you get when you call mosaic directly.
This means that we have to be carful with how we handle which arguments
are required, and generally need to ensure that a default value always
is applied for any required arguments.