Support --rolling with --end-date for statusduration.
The logic goes that we look for all issues that transitioned into the target
state before or during the query period, excluding any which transitioned out
of the state before it.

This required updates to the way the duration is calculated which assumed that
"today" was a suitable substitute for the end date.  The end of the query
period is used instead.