Allow excluding some issues by bounding duration.
This allows excluding outliers.

i.e., anything that passes through a state in 0.01 of a day, which likely
corresponds with someone forgetting to update the state of their card and
fast-tracking it through code review long after the code review is really done.

Or, i.e., anything that passes through a state in 200 days, which likely
corresponds with someone forgetting to close or move a card forwards long after
the action is actually done.